# Improve rate store refresh latency