--- a/pkg/ingester/stream_rate_calculator_test.go
+++ b/pkg/ingester/stream_rate_calculator_test.go
@@ -12,14 +12,16 @@ func TestStreamRateCalculator(t *testing.T) {
 	defer calc.Stop()
 
 	for i := 0; i < 100; i++ {
-		calc.Record(0, 100)
+		calc.Record(1, 1, 100)
 	}
 
 	require.Eventually(t, func() bool {
-		return calc.RateFor(0) == 10000
+		rates := calc.Rates()
+		return len(rates) > 0 && rates[0].Rate == 10000
 	}, 2*time.Second, 250*time.Millisecond)
 
 	require.Eventually(t, func() bool {
-		return calc.RateFor(0) == 0
+		rates := calc.Rates()
+		return len(rates) == 0
 	}, 2*time.Second, 250*time.Millisecond)
 }