Rather than iterating streams and getting rates per stream, keep all necessary information in the rate calculator and just hand back the current rates when asked.

This change took p99 latencies from ~1.8s to ~400ms and p50 latences from ~1s to 6ms.